Move test types in preparation snapshot tests for all of them
