+++ tests/Vogen.IntegrationTests/DiagnosticsTests/DisallowDefaultingTests.cs
