+++ tests/Vogen.IntegrationTests/DiagnosticsTests/DisallowNewTests.cs
