+++ tests/Vogen.IntegrationTests/SnapshotTests/Snapshots/ValueObjectGeneratorTests.No_namespace.verified.txt
