+++ tests/Vogen.IntegrationTests/SnapshotTests/Snapshots/ValueObjectGeneratorTests.Partial_struct_created_successfully.verified.txt
