+++ tests/Vogen.IntegrationTests/SnapshotTests/Snapshots/ValueObjectGeneratorTests.Produces_instances.verified.txt
