+++ tests/Vogen.IntegrationTests/SnapshotTests/Snapshots/ValueObjectGeneratorTests.Validation_with_PacalCased_validate_method.verified.txt
