+++ tests/Vogen.IntegrationTests/SnapshotTests/Snapshots/ValueObjectGeneratorTests.Validation_with_camelCased_validate_method.verified.txt
