--- tests/Vogen.IntegrationTests/Snapshots/ValueObjectGeneratorTests.Produces_instances.verified.txt
@@ -1,103 +1,0 @@
-﻿// ------------------------------------------------------------------------------
-// <auto-generated>
-//     This code was generated by a source generator named Vogen (https://github.com/SteveDunn/Vogen)
-//
-//     Changes to this file may cause incorrect behavior and will be lost if
-//     the code is regenerated.
-// </auto-generated>
-// ------------------------------------------------------------------------------
-
-using Vogen;
-
-namespace Whatever
-{
-
-    public partial struct CustomerId : System.IEquatable<CustomerId>
-    {
-        private readonly bool _isInitialized;
-        
-        private readonly int _value;
-
-        public readonly int Value
-        {
-            get
-            {
-                EnsureInitialized();
-                return _value;
-            }
-        }
-
-        public CustomerId()
-        {
-            throw new Vogen.ValueObjectValidationException("Validation skipped by attempting to use the default constructor. Please use the 'From' method for construction.");
-        }
-
-        private CustomerId(int value) 
-        {
-            _value = value;
-            _isInitialized = true;
-        }
-
-        public static CustomerId From(int value)
-        {
-            CustomerId instance = new CustomerId(value);
-
-            
-
-            return instance;
-        }
-
-        public readonly bool Equals(CustomerId other)
-        {
-            EnsureInitialized();
-            return System.Collections.Generic.EqualityComparer<int>.Default.Equals(Value, other.Value);
-        }
-
-        public readonly bool Equals(int primitive) => Value.Equals(primitive);
-
-        public readonly override bool Equals(object obj)
-        {
-            return obj is CustomerId && Equals((CustomerId) obj);
-        }
-
-        public static bool operator ==(CustomerId left, CustomerId right) => Equals(left, right);
-        public static bool operator !=(CustomerId left, CustomerId right) => !(left == right);
-
-        public static bool operator ==(CustomerId left, int right) => Equals(left.Value, right);
-        public static bool operator !=(CustomerId left, int right) => !Equals(left.Value, right);
-
-        public static bool operator ==(int left, CustomerId right) => Equals(left, right.Value);
-        public static bool operator !=(int left, CustomerId right) => !Equals(left, right.Value);
-
-        public readonly override int GetHashCode() => System.Collections.Generic.EqualityComparer<int>.Default.GetHashCode(_value);
-
-        public readonly override string ToString() => Value.ToString();
-
-        private readonly void EnsureInitialized()
-        {
-            if(!_isInitialized) throw new ValueObjectValidationException("Use of uninitialized Value Object.");
-        }
-
-        
-// instance...
-
-public static CustomerId Unspecified = new CustomerId(-1);
-
-// instance...
-
-public static CustomerId Unspecified1 = new CustomerId(-2);
-
-// instance...
-
-public static CustomerId Unspecified2 = new CustomerId(-3);
-
-// instance...
-
-public static CustomerId Unspecified3 = new CustomerId(-4);
-
-// instance...
-
-public static CustomerId Cust42 = new CustomerId(42);
-
-    }
-}
