+++ tests/Vogen.IntegrationTests/TestTypes/ByteVo.cs
