+++ tests/Vogen.IntegrationTests/TestTypes/CharVo.cs
