+++ tests/Vogen.IntegrationTests/TestTypes/DapperBoolVo.cs
