+++ tests/Vogen.IntegrationTests/TestTypes/DateTimeOffsetVo.cs
