+++ tests/Vogen.IntegrationTests/TestTypes/DateTimeVo.cs
