+++ tests/Vogen.IntegrationTests/TestTypes/DecimalVo.cs
