+++ tests/Vogen.IntegrationTests/TestTypes/DoubleVo.cs
