+++ tests/Vogen.IntegrationTests/TestTypes/FloatVo.cs
