+++ tests/Vogen.IntegrationTests/TestTypes/FooVo.cs
