+++ tests/Vogen.IntegrationTests/TestTypes/GuidVo.cs
