+++ tests/Vogen.IntegrationTests/TestTypes/IntoVo.cs
