+++ tests/Vogen.IntegrationTests/TestTypes/LongVo.cs
