+++ tests/Vogen.IntegrationTests/TestTypes/ShortVo.cs
