+++ tests/Vogen.IntegrationTests/TestTypes/StringVo.cs
