+++ tests/Vogen.IntegrationTests/Vogen.IntegrationTests.csproj
@@ -56,15 +56,11 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Snapshots\" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <None Update="DiagnosticsTests\Snapshots\ValueObjectGeneratorTests.Validation_with_camelCased_validate_method.verified.txt">
+	  <None Update="SnapshotTests\Snapshots\ValueObjectGeneratorTests.Validation_with_camelCased_validate_method.verified.txt">
 	    <ParentFile>$([System.String]::Copy('%(FileName)').Split('.')[0])</ParentFile>
 	    <DependentUpon>%(ParentFile).cs</DependentUpon>
 	  </None>
-	  <None Update="DiagnosticsTests\Snapshots\ValueObjectGeneratorTests.Validation_with_PacalCased_validate_method.verified.txt">
+	  <None Update="SnapshotTests\Snapshots\ValueObjectGeneratorTests.Validation_with_PacalCased_validate_method.verified.txt">
 	    <ParentFile>$([System.String]::Copy('%(FileName)').Split('.')[0])</ParentFile>
 	    <DependentUpon>%(ParentFile).cs</DependentUpon>
 	  </None>
